Skip merge commit if merge base isn't an ancestor
If the recorded merge base isn't an ancestor of the branch, it's really
misleading to make a merge commit claiming it is: it'll bring in a lot of other
apparent commits, but not include any of their contents. We still think we're
better off using the contents of the recorded base for 3-way merge, since it's
very likely that that commit was once an ancestor before a rebase, but there's
no valid git history to record.